+++ code/Folder.js
@@ -1,21 +1,14 @@
-import { FeaturesCollection } from "./Features"
-import { PLDBFile, runtimeCsvProps, isLanguage } from "./File"
-import {
-  FeatureSummary,
-  FolderInterface,
-  InverseRankings,
-  StringMap
-} from "./Interfaces"
-import { computeRankings } from "./Rankings"
-
-const { quickCache } = require("./quickCache")
 const path = require("path")
 const lodash = require("lodash")
+
 const { Table } = require("jtree/products/jtable.node.js")
 const { TreeNode } = require("jtree/products/TreeNode.js")
-const { Utils } = require("jtree/products/Utils.js")
 const { TreeBaseFolder } = require("jtree/products/treeBase.node.js")
 const { Disk } = require("jtree/products/Disk.node.js")
+const { Utils } = require("jtree/products/Utils.js")
+const { shiftRight, removeReturnChars } = Utils
+
+const { PLDBFile, isLanguage } = require("./File.js")
 
 const databaseFolder = path.join(__dirname, "..", "database")
 
@@ -35,8 +28,325 @@ const nodeToFlatObject = parentNode => {
   return newObject
 }
 
+const linkManyAftertext = links =>
+  links.map((link, index) => `${index + 1}.`).join(" ") + // notice the dot is part of the link. a hack to make it more unique for aftertext matching.
+  links.map((link, index) => `\n ${link} ${index + 1}.`).join("")
+
+// One feature maps to one grammar file that extends abstractFeatureNode
+class Feature {
+  constructor(node, collection) {
+    this.node = node
+    this.collection = collection
+    this.fileName = this.id + ".grammar"
+    this.id = node.id.replace("Node", "")
+  }
+
+  id = ""
+  fileName = ""
+
+  get permalink() {
+    return this.id + ".html"
+  }
+
+  previous
+  next
+
+  node
+  collection
+
+  get yes() {
+    return this.languagesWithThisFeature.length
+  }
+
+  get no() {
+    return this.languagesWithoutThisFeature.length
+  }
+
+  get percentage() {
+    const { yes, no } = this
+    const measurements = yes + no
+    return measurements < 100
+      ? "-"
+      : lodash.round((100 * yes) / measurements, 0) + "%"
+  }
+
+  get aka() {
+    return this.get("aka") // .join(" or "),
+  }
+
+  get token() {
+    return this.get("tokenKeyword")
+  }
+
+  get titleLink() {
+    return `../features/${this.permalink}`
+  }
+
+  get(word) {
+    return this.node.getFrom(`string ${word}`)
+  }
+
+  get title() {
+    return this.get("title") || this.id
+  }
+
+  get pseudoExample() {
+    return (this.get("pseudoExample") || "")
+      .replace(/\</g, "&lt;")
+      .replace(/\|/g, "&#124;")
+  }
+
+  get references() {
+    return (this.get("reference") || "").split(" ").filter(i => i)
+  }
+
+  get base() {
+    return this.collection.base
+  }
+
+  get languagesWithThisFeature() {
+    const { id } = this
+    return this.base
+      .getLanguagesWithFeatureResearched(id)
+      .filter(file => file.extendedFeaturesNode.get(id) === "true")
+  }
+
+  get languagesWithoutThisFeature() {
+    const { id } = this
+    return this.base
+      .getLanguagesWithFeatureResearched(id)
+      .filter(file => file.extendedFeaturesNode.get(id) === "false")
+  }
+
+  get summary() {
+    const {
+      id,
+      title,
+      fileName,
+      titleLink,
+      aka,
+      token,
+      yes,
+      no,
+      percentage,
+      pseudoExample
+    } = this
+    return {
+      id,
+      title,
+      fileName,
+      titleLink,
+      aka,
+      token,
+      yes,
+      no,
+      measurements: yes + no,
+      percentage,
+      pseudoExample
+    }
+  }
+
+  toScroll() {
+    const { title, id, fileName, references, previous, next } = this
+
+    const positives = this.languagesWithThisFeature
+    const positiveText = `* Languages *with* ${title} include ${positives
+      .map(file => `<a href="../languages/${file.permalink}">${file.title}</a>`)
+      .join(", ")}`
+
+    const negatives = this.languagesWithoutThisFeature
+    const negativeText = negatives.length
+      ? `* Languages *without* ${title} include ${negatives
+          .map(
+            file => `<a href="../languages/${file.permalink}">${file.title}</a>`
+          )
+          .join(", ")}`
+      : ""
+
+    const examples = positives
+      .filter(file => file.extendedFeaturesNode.getNode(id).length)
+      .map(file => {
+        return {
+          id: file.id,
+          title: file.title,
+          example: file.extendedFeaturesNode.getNode(id).childrenToString()
+        }
+      })
+    const grouped = lodash.groupBy(examples, "example")
+    const examplesText = Object.values(grouped)
+      .map(group => {
+        const links = group
+          .map(hit => `<a href="../languages/${hit.id}.html">${hit.title}</a>`)
+          .join(", ")
+        return `codeWithHeader Example from ${links}:
+ ${shiftRight(removeReturnChars(lodash.escape(group[0].example)), 1)}`
+      })
+      .join("\n\n")
+
+    let referencesText = ""
+    if (references.length)
+      referencesText = `* Read more about ${title} on the web: ${linkManyAftertext(
+        references
+      )}`
+
+    let descriptionText = ""
+    const description = this.node.get(`description`)
+    if (description) descriptionText = `* This question asks: ${description}`
+
+    return `import header.scroll
+
+title ${title}
+
+title ${title} - language feature
+ hidden
+
+html
+ <a class="prevLang" href="${previous.permalink}">&lt;</a>
+ <a class="nextLang" href="${next.permalink}">&gt;</a>
+
+viewSourceUrl https://github.com/breck7/pldb/blob/main/database/grammar/${fileName}
+
+startColumns 4
+
+${examplesText}
+
+${positiveText}
+
+${negativeText}
+
+${descriptionText}
+
+${referencesText}
+
+* HTML of this page generated by Features.ts
+ https://github.com/breck7/pldb/blob/main/code/Features.ts Features.ts
+
+endColumns
+
+keyboardNav ${previous.permalink} ${next.permalink}
+
+import ../footer.scroll
+`.replace(/\n\n\n+/g, "\n\n")
+  }
+}
+
+class FeaturesCollection {
+  base
+  features
+  constructor(base) {
+    this.base = base
+
+    const allGrammarNodes = Object.values(
+      base
+        .nodeAt(0)
+        .parsed.getDefinition()
+        ._getProgramNodeTypeDefinitionCache()
+    )
+
+    this.features = allGrammarNodes
+      .filter(node => node.get("extends") === "abstractFeatureNode")
+      .map(nodeDef => {
+        const feature = new Feature(nodeDef, this)
+        if (!feature.title) {
+          throw new Error(`Feature ${nodeDef.toString()} has no title.`)
+        }
+        return feature
+      })
+
+    let previous = this.features[this.features.length - 1]
+    this.features.forEach((feature, index) => {
+      feature.previous = previous
+      feature.next = this.features[index + 1]
+      previous = feature
+    })
+    this.features[this.features.length - 1].next = this.features[0]
+  }
+}
+
+const calcRanks = (folder, files) => {
+  const { inboundLinks } = folder
+  let objects = files.map(file => {
+    const id = file.id
+    const object = {}
+    object.id = id
+    object.jobs = folder.predictNumberOfJobs(file)
+    object.users = folder.predictNumberOfUsers(file)
+    object.facts = file.factCount
+    object.inboundLinks = inboundLinks[id].length
+    return object
+  })
+
+  objects = rankSort(objects, "jobs")
+  objects = rankSort(objects, "users")
+  objects = rankSort(objects, "facts")
+  objects = rankSort(objects, "inboundLinks")
+
+  objects.forEach((obj, rank) => {
+    // Drop the item this does the worst on, as it may be a flaw in PLDB.
+    const top3 = [
+      obj.jobsRank,
+      obj.usersRank,
+      obj.factsRank,
+      obj.inboundLinksRank
+    ]
+    obj.totalRank = lodash.sum(lodash.sortBy(top3).slice(0, 3))
+  })
+  objects = lodash.sortBy(objects, ["totalRank"])
+
+  const ranks = {}
+  objects.forEach((obj, index) => {
+    obj.index = index
+    ranks[obj.id] = obj
+  })
+  return ranks
+}
+
+const rankSort = (objects, key) => {
+  objects = lodash.sortBy(objects, [key])
+  objects.reverse()
+  let lastValue = objects[0][key]
+  let lastRank = 0
+  objects.forEach((obj, rank) => {
+    const theValue = obj[key]
+    if (lastValue === theValue) {
+      // A tie
+      obj[key + "Rank"] = lastRank
+    } else {
+      obj[key + "Rank"] = rank
+      lastRank = rank
+      lastValue = theValue
+    }
+  })
+  return objects
+}
+
+const computeRankings = folder => {
+  const ranks = calcRanks(folder, folder.getChildren())
+  const inverseRanks = makeInverseRanks(ranks)
+  const languageRanks = calcRanks(
+    folder,
+    folder.filter(file => file.isLanguage)
+  )
+  const inverseLanguageRanks = makeInverseRanks(languageRanks)
+
+  return {
+    ranks,
+    inverseRanks,
+    languageRanks,
+    inverseLanguageRanks
+  }
+}
+
+const makeInverseRanks = ranks => {
+  const inverseRanks = {}
+  Object.keys(ranks).forEach(id => {
+    inverseRanks[ranks[id].index] = ranks[id]
+  })
+  return inverseRanks
+}
+
 class PLDBFolder extends TreeBaseFolder {
-  static getBase(): PLDBFolder {
+  static getBase() {
     return new PLDBFolder()
       .setDir(path.join(databaseFolder, "things"))
       .setGrammarDir(path.join(databaseFolder, "grammar"))
@@ -46,8 +356,7 @@ class PLDBFolder extends TreeBaseFolder {
     return new TreeNode.Parser(PLDBFile)
   }
 
-  @quickCache
-  get inboundLinks(): StringMap {
+  get inboundLinks() {
     const inBoundLinks = {}
     this.forEach(file => (inBoundLinks[file.id] = []))
 
@@ -70,8 +379,7 @@ class PLDBFolder extends TreeBaseFolder {
     return new gpc().getDefinition()
   }
 
-  @quickCache
-  get searchIndex(): Map<string, string> {
+  get searchIndex() {
     const map = new Map()
     this.forEach(file => {
       const { id } = file
@@ -80,7 +388,7 @@ class PLDBFolder extends TreeBaseFolder {
     return map
   }
 
-  searchForEntity(query: string) {
+  searchForEntity(query) {
     if (query === undefined || query === "") return
     const { searchIndex } = this
     return (
@@ -90,15 +398,14 @@ class PLDBFolder extends TreeBaseFolder {
     )
   }
 
-  searchForEntityByFileExtensions(extensions: string[] = []) {
+  searchForEntityByFileExtensions(extensions = []) {
     const { extensionsMap } = this
     const hit = extensions.find(ext => extensionsMap.has(ext))
     return extensionsMap.get(hit)
   }
 
-  @quickCache
   get extensionsMap() {
-    const extensionsMap = new Map<string, string>()
+    const extensionsMap = new Map()
     this.topLanguages
       .slice(0)
       .reverse()
@@ -111,21 +418,20 @@ class PLDBFolder extends TreeBaseFolder {
     return extensionsMap
   }
 
-  getFile(id: string): PLDBFile | undefined {
+  getFile(id) {
     if (id === undefined) return undefined
     if (id.includes("/")) id = Disk.getFileName(id).replace(".pldb", "")
     return this.getNode(id)
   }
 
-  @quickCache
-  get topLanguages(): PLDBFile[] {
+  get topLanguages() {
     return lodash.sortBy(
       this.filter(lang => lang.isLanguage),
       "languageRank"
     )
   }
 
-  predictNumberOfUsers(file: PLDBFile) {
+  predictNumberOfUsers(file) {
     const mostRecents = [
       "linkedInSkill",
       "subreddit memberCount",
@@ -156,23 +462,21 @@ class PLDBFolder extends TreeBaseFolder {
     )
   }
 
-  predictNumberOfJobs(file: PLDBFile) {
+  predictNumberOfJobs(file) {
     return (
       Math.round(file.getMostRecentInt("linkedInSkill") * 0.01) +
       file.getMostRecentInt("indeedJobs")
     )
   }
 
-  @quickCache
   get rankings() {
     // Todo: once jtree is cleaned up, we should be able to remove this.
     // the problem is this class does implement FolderInterface, but Typescript doesn't know that
     // because it misses the inherited methods (filter and getChildren).
-    const folder: FolderInterface = <any>this
-    return computeRankings(folder)
+    return computeRankings(this)
   }
 
-  private _getFileAtRank(rank: number, ranks: InverseRankings) {
+  _getFileAtRank(rank, ranks) {
     const count = Object.keys(ranks).length
     if (rank < 0) rank = count - 1
     if (rank >= count) rank = 0
@@ -180,7 +484,7 @@ class PLDBFolder extends TreeBaseFolder {
   }
 
   _featureCache = {}
-  getLanguagesWithFeatureResearched(id: string) {
+  getLanguagesWithFeatureResearched(id) {
     if (!this._featureCache[id])
       this._featureCache[id] = this.topLanguages.filter(file =>
         file.extendedFeaturesNode.has(id)
@@ -188,52 +492,50 @@ class PLDBFolder extends TreeBaseFolder {
     return this._featureCache[id]
   }
 
-  @quickCache
-  get featuresMap(): Map<string, FeatureSummary> {
-    const featuresMap = new Map<string, FeatureSummary>()
+  get featuresMap() {
+    const featuresMap = new Map()
     this.topFeatures.forEach(feature => featuresMap.set(feature.id, feature))
     return featuresMap
   }
 
-  @quickCache get features() {
+  get features() {
     return new FeaturesCollection(this).features
   }
 
-  @quickCache
-  get topFeatures(): FeatureSummary[] {
+  get topFeatures() {
     const { features } = this
     const sorted = lodash.sortBy(features, "yes")
     sorted.reverse()
     return sorted
   }
 
-  getFileAtLanguageRank(rank: number) {
+  getFileAtLanguageRank(rank) {
     return this._getFileAtRank(rank, this.rankings.inverseLanguageRanks)
   }
 
-  getFileAtRank(rank: number) {
+  getFileAtRank(rank) {
     return this._getFileAtRank(rank, this.rankings.inverseRanks)
   }
 
-  predictPercentile(file: PLDBFile) {
+  predictPercentile(file) {
     const files = this.getChildren()
     const { ranks } = this.rankings
     return ranks[file.id].index / files.length
   }
 
-  getLanguageRankExplanation(file: PLDBFile) {
+  getLanguageRankExplanation(file) {
     return this.rankings.languageRanks[file.id]
   }
 
-  getLanguageRank(file: PLDBFile) {
+  getLanguageRank(file) {
     return this.rankings.languageRanks[file.id].index
   }
 
-  getRank(file: PLDBFile) {
+  getRank(file) {
     return this.rankings.ranks[file.id].index
   }
 
-  makeId(title: string) {
+  makeId(title) {
     let id = Utils.titleToPermalink(title)
     let newId = id
     if (!this.getFile(newId)) return newId
@@ -246,7 +548,7 @@ class PLDBFolder extends TreeBaseFolder {
     )
   }
 
-  createFile(content: string, id?: string) {
+  createFile(content, id) {
     if (id === undefined) {
       const title = new TreeNode(content).get("title")
       if (!title)
@@ -270,12 +572,11 @@ class PLDBFolder extends TreeBaseFolder {
   }
 
   get exampleCounts() {
-    const counts = new Map<string, number>()
+    const counts = new Map()
     this.forEach(file => counts.set(file.id, file.exampleCount))
     return counts
   }
 
-  @quickCache
   get colNameToGrammarDefMap() {
     const map = new Map()
     this.nodesForCsv.forEach(node => {
@@ -291,11 +592,7 @@ class PLDBFolder extends TreeBaseFolder {
     return this.columnDocumentation.map(col => col.Column)
   }
 
-  groupByListValues(
-    listColumnName: string,
-    files = this.files,
-    delimiter = " && "
-  ) {
+  groupByListValues(listColumnName, files = this.files, delimiter = " && ") {
     const values = {}
     files.forEach(file => {
       const value = file.get(listColumnName)
@@ -309,7 +606,7 @@ class PLDBFolder extends TreeBaseFolder {
   }
 
   // todo: is there already a way to do this in jtree?
-  getFilePathAndLineNumberWhereGrammarNodeIsDefined(nodeTypeId: string) {
+  getFilePathAndLineNumberWhereGrammarNodeIsDefined(nodeTypeId) {
     const { grammarFileMap } = this
     const regex = new RegExp(`^${nodeTypeId}`, "gm")
     let filePath
@@ -325,7 +622,6 @@ class PLDBFolder extends TreeBaseFolder {
     return { filePath, lineNumber }
   }
 
-  @quickCache
   get grammarFileMap() {
     const map = {}
     this.grammarFilePaths.forEach(
@@ -334,7 +630,6 @@ class PLDBFolder extends TreeBaseFolder {
     return map
   }
 
-  @quickCache
   get columnDocumentation() {
     // Return columns with documentation sorted in the most interesting order.
 
@@ -436,9 +731,10 @@ wikipedia`.split("\n")
     return sortedCols
   }
 
-  @quickCache
   get nodesForCsv() {
-    const runTimeProps = Object.keys(runtimeCsvProps)
+    const runTimeProps = `pldbId bookCount paperCount hopl exampleCount numberOfUsers numberOfRepos numberOfJobs languageRank rank factCount lastActivity`.split(
+      " "
+    )
     return this.map(file => {
       const clone = file.parsed.clone()
       clone.getTopDownArray().forEach(node => {
@@ -455,17 +751,14 @@ wikipedia`.split("\n")
     })
   }
 
-  @quickCache
   get typedMapJson() {
     return JSON.stringify(this.typedMap, null, 2)
   }
 
-  @quickCache
   get keywordsOneHotCsv() {
     return new TreeNode(this.keywordsOneHot).toCsv()
   }
 
-  @quickCache
   get searchIndexJson() {
     return JSON.stringify(
       this.objectsForCsv.map(object => {
@@ -481,14 +774,12 @@ wikipedia`.split("\n")
     )
   }
 
-  @quickCache
   get objectsForCsv() {
     return lodash.sortBy(this.nodesForCsv.map(nodeToFlatObject), item =>
       parseInt(item.rank)
     )
   }
 
-  @quickCache
   get csvBuildOutput() {
     const { colNamesForCsv, objectsForCsv, columnDocumentation } = this
 
@@ -539,7 +830,6 @@ wikipedia`.split("\n")
     return sources.sort()
   }
 
-  @quickCache
   get keywordsOneHot() {
     const { keywordsTable } = this
     const allKeywords = keywordsTable.rows.map(row => row.keyword)
@@ -560,22 +850,18 @@ wikipedia`.split("\n")
     return rows
   }
 
-  @quickCache
   get bytes() {
     return this.toString().length
   }
 
-  @quickCache
   get cachedErrors() {
     return this.errors
   }
 
-  @quickCache
   get factCount() {
     return lodash.sum(this.map(file => file.factCount))
   }
 
-  @quickCache
   get keywordsTable() {
     const langsWithKeywords = this.topLanguages.filter(file =>
       file.has("keywords")
@@ -600,7 +886,7 @@ wikipedia`.split("\n")
     })
 
     const rows = Object.values(keywordsMap)
-    rows.forEach((row: any) => {
+    rows.forEach(row => {
       row.count = row.ids.length
       row.langs = row.ids
         .map(id => {
@@ -620,4 +906,4 @@ wikipedia`.split("\n")
   }
 }
 
-export { PLDBFolder }
+module.exports = { PLDBFolder }
