+++ code/crawlers/package.json
@@ -22,6 +22,10 @@
     "superagent": "^5.3.1",
     "whois-json": "^2.0.4",
     "wikijs": "^6.3.3",
+    "typescript": ">=4.9.5",
     "yaml": "^2.1.1"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.18"
   }
 }
