--- code/dockerBuildInstructions.scroll
@@ -1,19 +1,0 @@
-importOnly
-## To build the full site in docker
-code
- # -- Run PLDB in docker
- docker run -it -d -v /c/AndelaWork/pldb:/app -p 4444:4444 --name pldbnode node:16 bash
- # -- into docker interactive session run the follow:
- git clone https://github.com/breck7/pldb
- cd pldb
- npm install .
- npm run tsc
- npm i -g cloc
- npm run build
- # (Optional run TESTS):
-  npm run test
- # (Following updates):
- npm run format
- npm run local
- # navigate site in -->
- http://localhost:4444/index.html
