--- code/quickCache.js
@@ -1,14 +1,0 @@
-// Memoization for immutable getters. Run the function once for this instance and cache the result.
-const quickCache = (target, propName, descriptor) => {
-  const originalFn = descriptor.get
-  descriptor.get = function() {
-    if (!this.quickCacheCache) this.quickCacheCache = {}
-    const quickCacheCache = this.quickCacheCache
-    if (quickCacheCache[propName] === undefined)
-      quickCacheCache[propName] = originalFn.apply(this)
-    return quickCacheCache[propName]
-  }
-}
-const clearQuickCache = obj => delete obj.quickCacheCache
-
-module.exports = { quickCache, clearQuickCache }
