--- code/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "types": ["node"],
-    "lib": ["es2017", "dom"],
-    "module": "commonjs",
-    "moduleResolution": "Node",
-    "target": "es2017",
-    "experimentalDecorators": true
-  },
-  "exclude": ["**/ignore", "**/cache"]
-}
