+++ package.json
@@ -2,20 +2,18 @@
   "name": "pldb",
   "description": "PLDB: a Programming Language Database",
   "scripts": {
-    "tsc": "cd code; tsc; cd ..; npm run makeCrawlersExecutable",
     "test": "./code/test.js",
-    "search": "./code/cliSearch.js",
     "local": "./code/PLDBServer.js startDevServer",
+    "search": "./code/cliSearch.js",
     "format": "./code/PLDBServer.js formatDatabase",
+    "tsc": "cd code/crawlers; tsc; chmod u+x *.js; chmod u+x **/*.js; chmod u+x code/**/**/*.js",
     "crawl": "./code/crawlers/Crawler.js crawl",
     "githubs": "./code/crawlers/github.com/GitHub.js fetchAllRepoData && ./code/crawlers/github.com/GitHub.js writeAllRepoData",
     "fetchContributors": "wget -O site/pages/contributors.json https://api.github.com/repos/breck7/pldb/contributors",
-    "clean": "cd code; tsc --build --clean",
     "cloc": "cloc --vcs git . --read-lang-def=code/clocLangs.txt",
     "checkLinks": "linkinator -r https://pldb.com > ignore/brokenLinks.txt",
     "jtreeBeta": "npm install ~/jtree",
     "scrollBeta": "npm install ~/scroll",
-    "makeCrawlersExecutable": "chmod u+x code/**/*.js; chmod u+x code/**/*.js; chmod u+x code/**/**/*.js",
     "publish": "cd code; npm publish",
     "deploy": "ssh pldb.com; git pull; x tsc; x restart",
     "fetchLogs": "scp file pldb.com:/root/pldb/ignore/treeBaseServerLog.tree ignore/prodLogs/",
@@ -52,10 +50,6 @@
     "scroll-cli": "^58.0.1",
     "simple-git": "^3.16.0",
     "simple-statistics": "latest",
-    "typescript": ">=4.9.5"
-  },
-  "devDependencies": {
-    "@types/node": "^18.11.18",
     "tap": "^16.3.4"
   }
 }
